Extract FeedbackManager into a new file
